Improve switching between date/time parameters
Now when switching between date/time parameters when setting them it will skip deliminator like colon or slash.
Now when setting month it allows only to choose second digit. Changing the second digit will change the first when it should (e.g. 09 to 10 if increased; 11 to 10, 10 to 09 if decreased). This is not yet implemented. The next commit will fix setting month and add this